+++ lib/address_util.dart
@@ -1,7 +1,7 @@
-import 'package:uniris_lib_dart/src/api.dart';
-import 'package:uniris_lib_dart/src/crypto.dart' as crypto;
-import 'package:uniris_lib_dart/src/model/key_pair.dart';
-import 'package:uniris_lib_dart/src/utils.dart';
+import 'package:uniris_lib_dart/api.dart';
+import 'package:uniris_lib_dart/crypto.dart' as crypto;
+import 'package:uniris_lib_dart/model/key_pair.dart';
+import 'package:uniris_lib_dart/utils.dart';
 
 /*
  * Derive an address