+++ lib/transaction_builder.dart
@@ -4,9 +4,9 @@
 
 import 'dart:convert' show json, utf8, jsonEncode;
 import 'dart:typed_data' show Uint8List;
-import 'package:uniris_lib_dart/src/model/key_pair.dart';
-import 'package:uniris_lib_dart/src/utils.dart';
-import 'package:uniris_lib_dart/src/crypto.dart' as crypto;
+import 'package:uniris_lib_dart/model/key_pair.dart';
+import 'package:uniris_lib_dart/utils.dart';
+import 'package:uniris_lib_dart/crypto.dart' as crypto;
 
 const Map<String, int> txTypes = {
   "identity": 0,